feat: Refactor order confirmation and payment success pages to use wrapper components
